requirements.txt
# All Python dependencies (pinned versions)

fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
anthropic>=0.40.0
PyMuPDF==1.24.11
python-docx==1.1.2
pytesseract==0.3.13
Pillow==10.4.0
python-dotenv==1.0.1
pydantic==2.8.2
httpx>=0.27.0
google-generativeai
